refactor: :recycle: hook for themes
adding a hook to manage the themes and have the code organized
